Rename action_router to block_actions_router and suggestion_router to block_suggestion_router.
