+++ src/functions/interactivity/block_suggestion_router_test.ts
@@ -5,9 +5,8 @@ import {
   assertRejects,
   mock,
 } from "../../dev_deps.ts";
-import { BlockActionsRouter } from "./action_router.ts";
-import type { ActionContext } from "./types.ts";
-import type { BlockAction } from "./block_kit_types.ts";
+import { BlockSuggestionRouter } from "./block_suggestion_router.ts";
+import type { SuggestionContext } from "./types.ts";
 import type {
   FunctionParameters,
   FunctionRuntimeParameters,
@@ -22,35 +21,37 @@ import { DefineFunction, Schema } from "../../mod.ts";
 // Helper test types
 // TODO: maybe we want to export this for userland usage at some point?
 // Very much a direct copy from the existing main function tester types and utilties in src/functions/tester
-type SlackActionHandlerTesterArgs<InputParameters extends FunctionParameters> =
+type SlackSuggestionHandlerTesterArgs<
+  InputParameters extends FunctionParameters,
+> =
   & Partial<
-    ActionContext<InputParameters>
+    SuggestionContext<InputParameters>
   >
   & {
     inputs: InputParameters;
   };
 
-type CreateActionHandlerContext<
+type CreateSuggestionHandlerContext<
   InputParameters extends ParameterSetDefinition,
   RequiredInput extends PossibleParameterKeys<InputParameters>,
 > = {
   (
-    args: SlackActionHandlerTesterArgs<
+    args: SlackSuggestionHandlerTesterArgs<
       FunctionRuntimeParameters<InputParameters, RequiredInput>
     >,
-  ): ActionContext<
+  ): SuggestionContext<
     FunctionRuntimeParameters<InputParameters, RequiredInput>
   >;
 };
 
-type SlackActionHandlerTesterResponse<
+type SlackSuggestionHandlerTesterResponse<
   InputParameters extends ParameterSetDefinition,
   RequiredInput extends PossibleParameterKeys<InputParameters>,
 > = {
-  createContext: CreateActionHandlerContext<InputParameters, RequiredInput>;
+  createContext: CreateSuggestionHandlerContext<InputParameters, RequiredInput>;
 };
 
-type SlackActionHandlerTesterFn = {
+type SlackSuggestionHandlerTesterFn = {
   // Accept a Slack Function
   <
     InputParameters extends ParameterSetDefinition,
@@ -64,21 +65,46 @@ type SlackActionHandlerTesterFn = {
       RequiredInput,
       RequiredOutput
     >,
-  ): SlackActionHandlerTesterResponse<
+  ): SlackSuggestionHandlerTesterResponse<
     InputParameters,
     RequiredInput
   >;
 };
-// Helper test fixtures and utilities
-const DEFAULT_ACTION: BlockAction = {
-  type: "button",
-  block_id: "block_id",
-  action_ts: `${new Date().getTime()}`,
-  action_id: "action_id",
-  text: { type: "plain_text", text: "duncare", emoji: false },
-  style: "danger",
-};
-const SlackActionHandlerTester: SlackActionHandlerTesterFn = <
+
+const DEFAULT_ACTION_ID = "action_id";
+const DEFAULT_BLOCK_ID = "block_id";
+// deno-lint-ignore no-explicit-any
+const generateSuggestion = (inputs: any) => ({
+  block_id: DEFAULT_BLOCK_ID,
+  action_id: DEFAULT_ACTION_ID,
+  value: "ohai",
+  type: "block_suggestion",
+  function_data: {
+    execution_id: "123",
+    function: { callback_id: "456" },
+    inputs,
+  },
+  interactivity: {
+    interactor: {
+      secret: "shhhh",
+      id: "123",
+    },
+    interactivity_pointer: "123.asdf",
+  },
+  user: {
+    id: "123",
+    name: "asdf",
+    team_id: "123",
+  },
+  team: {
+    id: "123",
+    domain: "asdf",
+  },
+  enterprise: null,
+  api_app_id: "123",
+});
+
+const SlackSuggestionHandlerTester: SlackSuggestionHandlerTesterFn = <
   InputParameters extends ParameterSetDefinition,
   OutputParameters extends ParameterSetDefinition,
   RequiredInput extends PossibleParameterKeys<InputParameters>,
@@ -91,7 +117,7 @@ const SlackActionHandlerTester: SlackActionHandlerTesterFn = <
     RequiredOutput
   >,
 ) => {
-  const createContext: CreateActionHandlerContext<
+  const createContext: CreateSuggestionHandlerContext<
     InputParameters,
     RequiredInput
   > = (
@@ -101,37 +127,6 @@ const SlackActionHandlerTester: SlackActionHandlerTesterFn = <
       InputParameters,
       RequiredInput
     >;
-    const DEFAULT_BODY = {
-      type: "block_actions",
-      actions: [DEFAULT_ACTION],
-      function_data: {
-        execution_id: "123",
-        function: { callback_id: "456" },
-        inputs,
-      },
-      interactivity: {
-        interactor: {
-          secret: "shhhh",
-          id: "123",
-        },
-        interactivity_pointer: "123.asdf",
-      },
-      user: {
-        id: "123",
-        name: "asdf",
-        team_id: "123",
-      },
-      team: {
-        id: "123",
-        domain: "asdf",
-      },
-      enterprise: null,
-      is_enterprise_install: false,
-      api_app_id: "123",
-      token: "123",
-      trigger_id: "123",
-      response_url: "asdf",
-    };
     const token = args.token || "slack-function-test-token";
 
     return {
@@ -141,8 +136,7 @@ const SlackActionHandlerTester: SlackActionHandlerTesterFn = <
       client: SlackAPI(token),
       team_id: args.team_id || "test-team-id",
       enterprise_id: "",
-      action: args.action || DEFAULT_ACTION,
-      body: args.body || DEFAULT_BODY,
+      body: args.body || generateSuggestion(inputs),
     };
   };
 
@@ -161,75 +155,77 @@ const func = DefineFunction({
     required: ["garbage"],
   },
 });
-const { createContext } = SlackActionHandlerTester(func);
+const { createContext } = SlackSuggestionHandlerTester(func);
 const inputs = { garbage: "in, garbage out" };
 
 const getRouter = () => {
-  return BlockActionsRouter(func);
+  return BlockSuggestionRouter(func);
 };
 
-Deno.test("ActionsRouter", async (t) => {
+Deno.test("SuggestionRouter", async (t) => {
   await t.step(
-    "export method returns result of handler when matching action comes in and baseline handler context parameters are present and exist",
+    "export method returns result of handler when matching suggestion comes in and baseline handler context parameters are present and exist",
     async () => {
       const router = getRouter();
       let handlerCalled = false;
-      router.addHandler(DEFAULT_ACTION.action_id, (ctx) => {
+      router.addHandler(DEFAULT_ACTION_ID, (ctx) => {
         assertExists(ctx.inputs);
         assertEquals<string>(ctx.inputs.garbage, inputs.garbage);
         assertExists(ctx.token);
-        assertExists(ctx.action);
+        assertExists(ctx.body);
         assertExists(ctx.env);
         assertExists(ctx.client);
         handlerCalled = true;
+        return { options: [] };
       });
       await router(createContext({ inputs }));
-      assertEquals(handlerCalled, true, "action handler not called!");
+      assertEquals(handlerCalled, true, "suggestion handler not called!");
     },
   );
 });
 
-Deno.test("ActionsRouter action matching happy path", async (t) => {
+Deno.test("SuggestionRouter matching happy path", async (t) => {
   await t.step("simple string matching to action_id", async () => {
     const router = getRouter();
     let handlerCalled = false;
-    router.addHandler(DEFAULT_ACTION.action_id, (ctx) => {
+    router.addHandler(DEFAULT_ACTION_ID, (ctx) => {
       assertExists(ctx.inputs);
       assertExists(ctx.token);
-      assertExists(ctx.action);
+      assertExists(ctx.body);
       assertExists(ctx.env);
       assertExists(ctx.client);
       handlerCalled = true;
+      return { options: [] };
     });
     await router(createContext({ inputs }));
-    assertEquals(handlerCalled, true, "action handler not called!");
+    assertEquals(handlerCalled, true, "suggestion handler not called!");
   });
   await t.step("array of strings matching to action_id", async () => {
     const router = getRouter();
-    const handler = mock.spy();
-    router.addHandler(["nope", DEFAULT_ACTION.action_id], handler);
+    const handler = mock.spy(() => ({ options: [] }));
+    router.addHandler(["nope", DEFAULT_ACTION_ID], handler);
     await router(createContext({ inputs }));
     mock.assertSpyCalls(handler, 1);
   });
   await t.step("regex matching to action_id", async () => {
     const router = getRouter();
-    const handler = mock.spy();
+    const handler = mock.spy(() => ({ options: [] }));
     router.addHandler(/action/, handler);
     await router(createContext({ inputs }));
     mock.assertSpyCalls(handler, 1);
   });
   await t.step("{action_id:string} matching to action_id", async () => {
     const router = getRouter();
-    const handler = mock.spy();
-    router.addHandler({ action_id: DEFAULT_ACTION.action_id }, handler);
+    const handler = mock.spy(() => ({ options: [] }));
+    router.addHandler({ action_id: DEFAULT_ACTION_ID }, handler);
     await router(createContext({ inputs }));
     mock.assertSpyCalls(handler, 1);
   });
   await t.step("{action_id:[string]} matching to action_id", async () => {
     const router = getRouter();
-    const handler = mock.spy();
+    const handler = mock.spy(() => ({ options: [] }));
     router.addHandler(
-      { action_id: ["hahtrickedyou", DEFAULT_ACTION.action_id] },
+      { action_id: ["hahtrickedyou", DEFAULT_ACTION_ID] },
       handler,
     );
     await router(createContext({ inputs }));
@@ -237,23 +233,23 @@ Deno.test("ActionsRouter action matching happy path", async (t) => {
   });
   await t.step("{action_id:regex} matching to action_id", async () => {
     const router = getRouter();
-    const handler = mock.spy();
+    const handler = mock.spy(() => ({ options: [] }));
     router.addHandler({ action_id: /action/ }, handler);
     await router(createContext({ inputs }));
     mock.assertSpyCalls(handler, 1);
   });
   await t.step("{block_id:string} matching to block_id", async () => {
     const router = getRouter();
-    const handler = mock.spy();
-    router.addHandler({ block_id: DEFAULT_ACTION.block_id }, handler);
+    const handler = mock.spy(() => ({ options: [] }));
+    router.addHandler({ block_id: DEFAULT_BLOCK_ID }, handler);
     await router(createContext({ inputs }));
     mock.assertSpyCalls(handler, 1);
   });
   await t.step("{block_id:[string]} matching to block_id", async () => {
     const router = getRouter();
-    const handler = mock.spy();
+    const handler = mock.spy(() => ({ options: [] }));
     router.addHandler(
-      { block_id: ["lol", DEFAULT_ACTION.block_id] },
+      { block_id: ["lol", DEFAULT_BLOCK_ID] },
       handler,
     );
     await router(createContext({ inputs }));
@@ -261,7 +257,7 @@ Deno.test("ActionsRouter action matching happy path", async (t) => {
   });
   await t.step("{block_id:regex} matching to block_id", async () => {
     const router = getRouter();
-    const handler = mock.spy();
+    const handler = mock.spy(() => ({ options: [] }));
     router.addHandler({ block_id: /block/ }, handler);
     await router(createContext({ inputs }));
     mock.assertSpyCalls(handler, 1);
@@ -270,10 +266,10 @@ Deno.test("ActionsRouter action matching happy path", async (t) => {
     "{block_id:string, action_id:string} matching to both block_id and action_id",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
-        block_id: DEFAULT_ACTION.block_id,
-        action_id: DEFAULT_ACTION.action_id,
+        block_id: DEFAULT_BLOCK_ID,
+        action_id: DEFAULT_ACTION_ID,
       }, handler);
       await router(createContext({ inputs }));
       mock.assertSpyCalls(handler, 1);
@@ -283,10 +279,10 @@ Deno.test("ActionsRouter action matching happy path", async (t) => {
     "{block_id:string, action_id:[string]} matching to both block_id and action_id",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
-        block_id: DEFAULT_ACTION.block_id,
-        action_id: ["notthisoneeither", DEFAULT_ACTION.action_id],
+        block_id: DEFAULT_BLOCK_ID,
+        action_id: ["notthisoneeither", DEFAULT_ACTION_ID],
       }, handler);
       await router(createContext({ inputs }));
       mock.assertSpyCalls(handler, 1);
@@ -296,9 +292,9 @@ Deno.test("ActionsRouter action matching happy path", async (t) => {
     "{block_id:string, action_id:regex} matching to both block_id and action_id",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler(
-        { block_id: DEFAULT_ACTION.block_id, action_id: /action/ },
+        { block_id: DEFAULT_BLOCK_ID, action_id: /action/ },
         handler,
       );
       await router(createContext({ inputs }));
@@ -309,10 +305,10 @@ Deno.test("ActionsRouter action matching happy path", async (t) => {
     "{block_id:[string], action_id:string} matching to both block_id and action_id",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
-        block_id: ["notthistime", DEFAULT_ACTION.block_id],
-        action_id: DEFAULT_ACTION.action_id,
+        block_id: ["notthistime", DEFAULT_BLOCK_ID],
+        action_id: DEFAULT_ACTION_ID,
       }, handler);
       await router(createContext({ inputs }));
       mock.assertSpyCalls(handler, 1);
@@ -322,10 +318,10 @@ Deno.test("ActionsRouter action matching happy path", async (t) => {
     "{block_id:[string], action_id:[string]} matching to both block_id and action_id",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
-        block_id: ["notthistime", DEFAULT_ACTION.block_id],
-        action_id: ["gotyougood", DEFAULT_ACTION.action_id],
+        block_id: ["notthistime", DEFAULT_BLOCK_ID],
+        action_id: ["gotyougood", DEFAULT_ACTION_ID],
       }, handler);
       await router(createContext({ inputs }));
       mock.assertSpyCalls(handler, 1);
@@ -335,9 +331,9 @@ Deno.test("ActionsRouter action matching happy path", async (t) => {
     "{block_id:[string], action_id:regex} matching to both block_id and action_id",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
-        block_id: ["notthistime", DEFAULT_ACTION.block_id],
+        block_id: ["notthistime", DEFAULT_BLOCK_ID],
         action_id: /action/,
       }, handler);
       await router(createContext({ inputs }));
@@ -348,9 +344,9 @@ Deno.test("ActionsRouter action matching happy path", async (t) => {
     "{block_id:regex, action_id:string} matching to both block_id and action_id",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler(
-        { block_id: /block/, action_id: DEFAULT_ACTION.action_id },
+        { block_id: /block/, action_id: DEFAULT_ACTION_ID },
         handler,
       );
       await router(createContext({ inputs }));
@@ -361,10 +357,10 @@ Deno.test("ActionsRouter action matching happy path", async (t) => {
     "{block_id:regex, action_id:[string]} matching to both block_id and action_id",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         block_id: /block/,
-        action_id: ["hahanope", DEFAULT_ACTION.action_id],
+        action_id: ["hahanope", DEFAULT_ACTION_ID],
       }, handler);
       await router(createContext({ inputs }));
       mock.assertSpyCalls(handler, 1);
@@ -374,7 +370,7 @@ Deno.test("ActionsRouter action matching happy path", async (t) => {
     "{block_id:regex, action_id:regex} matching to both block_id and action_id",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({ block_id: /block/, action_id: /action/ }, handler);
       await router(createContext({ inputs }));
       mock.assertSpyCalls(handler, 1);
@@ -382,8 +378,8 @@ Deno.test("ActionsRouter action matching happy path", async (t) => {
   );
 });
 
-Deno.test("ActionsRouter action matching sad path", async (t) => {
-  await t.step("unhandled action should throw", async () => {
+Deno.test("SuggestionRouter matching sad path", async (t) => {
+  await t.step("unhandled suggestion should throw", async () => {
     const router = getRouter();
     await assertRejects(() => router(createContext({ inputs })));
   });
@@ -393,7 +389,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
       "not matching action_id: string",
       async () => {
         const router = getRouter();
-        const handler = mock.spy();
+        const handler = mock.spy(() => ({ options: [] }));
         router.addHandler("nope", handler);
 
         await assertRejects(() => router(createContext({ inputs })));
@@ -406,7 +402,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching action_id: string[]",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler(["nope", "nuh uh"], handler);
 
       await assertRejects(() => router(createContext({ inputs })));
@@ -418,7 +414,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching action_id: regex",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler(/regex/, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
@@ -429,8 +425,8 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string}",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
-      router.addHandler({ action_id: "nope" }, () => handler);
+      const handler = mock.spy(() => ({ options: [] }));
+      router.addHandler({ action_id: "nope" }, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
     },
@@ -439,7 +435,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string[]}",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({ action_id: ["nope", "nuh uh"] }, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
@@ -449,7 +445,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: regex}",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({ action_id: /regex/ }, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
@@ -459,7 +455,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {block_id: string}",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({ block_id: "nope" }, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
@@ -469,7 +465,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {block_id: string[]}",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({ block_id: ["nope", "nuh uh"] }, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
@@ -479,7 +475,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {block_id: regex}",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({ block_id: /regex/ }, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
@@ -489,7 +485,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string, block_id: regex}, action_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         action_id: "not good enough",
         block_id: /block/,
@@ -502,9 +498,9 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string, block_id: regex}, block_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
-        action_id: "action_id",
+        action_id: DEFAULT_ACTION_ID,
         block_id: /noway/,
       }, handler);
       await assertRejects(() => router(createContext({ inputs })));
@@ -515,10 +511,10 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string, block_id: string[]}, action_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         action_id: "not good enough",
-        block_id: ["notthisonebut", "block_id"],
+        block_id: ["notthisonebut", DEFAULT_BLOCK_ID],
       }, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
@@ -528,9 +524,9 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string, block_id: string}, block_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
-        action_id: "action_id",
+        action_id: DEFAULT_ACTION_ID,
         block_id: ["this", "wont", "work"],
       }, handler);
       await assertRejects(() => router(createContext({ inputs })));
@@ -541,10 +537,10 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string, block_id: string}, action_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         action_id: "not good enough",
-        block_id: "block_id",
+        block_id: DEFAULT_BLOCK_ID,
       }, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
@@ -554,9 +550,9 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string, block_id: string}, block_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
-        action_id: "action_id",
+        action_id: DEFAULT_ACTION_ID,
         block_id: "nicetry",
       }, handler);
       await assertRejects(() => router(createContext({ inputs })));
@@ -567,7 +563,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string[], block_id: regex}, action_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         action_id: ["not", "good", "enough"],
         block_id: /block/,
@@ -580,9 +576,9 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string[], block_id: regex}, block_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
-        action_id: ["decoy", "action_id"],
+        action_id: ["decoy", DEFAULT_ACTION_ID],
         block_id: /noway/,
       }, handler);
       await assertRejects(() => router(createContext({ inputs })));
@@ -593,10 +589,10 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string[], block_id: string[]}, action_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         action_id: ["not", "good", "enough"],
-        block_id: ["notthisonebut", "block_id"],
+        block_id: ["notthisonebut", DEFAULT_BLOCK_ID],
       }, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
@@ -606,9 +602,9 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string[], block_id: string}, block_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
-        action_id: ["decoy", "action_id"],
+        action_id: ["decoy", DEFAULT_ACTION_ID],
         block_id: ["this", "wont", "work"],
       }, handler);
       await assertRejects(() => router(createContext({ inputs })));
@@ -619,10 +615,10 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string[], block_id: string}, action_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         action_id: ["not", "good", "enough"],
-        block_id: "block_id",
+        block_id: DEFAULT_BLOCK_ID,
       }, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
@@ -632,9 +628,9 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: string[], block_id: string}, block_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
-        action_id: ["decoy", "action_id"],
+        action_id: ["decoy", DEFAULT_ACTION_ID],
         block_id: "nicetry",
       }, handler);
       await assertRejects(() => router(createContext({ inputs })));
@@ -645,7 +641,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: regex, block_id: regex}, action_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         action_id: /heh/,
         block_id: /block/,
@@ -658,7 +654,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: regex, block_id: regex}, block_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         action_id: /action/,
         block_id: /noway/,
@@ -671,10 +667,10 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: regex, block_id: string[]}, action_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         action_id: /hah/,
-        block_id: ["notthisonebut", "block_id"],
+        block_id: ["notthisonebut", DEFAULT_BLOCK_ID],
       }, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
@@ -684,7 +680,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: regex, block_id: string}, block_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         action_id: /action/,
         block_id: ["this", "wont", "work"],
@@ -697,10 +693,10 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: regex, block_id: string}, action_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         action_id: /huh/,
-        block_id: "block_id",
+        block_id: DEFAULT_BLOCK_ID,
       }, handler);
       await assertRejects(() => router(createContext({ inputs })));
       mock.assertSpyCalls(handler, 0);
@@ -710,7 +706,7 @@ Deno.test("ActionsRouter action matching sad path", async (t) => {
     "not matching {action_id: regex, block_id: string}, block_id does not match",
     async () => {
       const router = getRouter();
-      const handler = mock.spy();
+      const handler = mock.spy(() => ({ options: [] }));
       router.addHandler({
         action_id: /action/,
         block_id: "nicetry",
