+++ src/functions/interactivity/mod.ts
@@ -1,3 +1,3 @@
-export { BlockActionsRouter } from "./action_router.ts";
+export { BlockActionsRouter } from "./block_actions_router.ts";
 export { ViewsRouter } from "./view_router.ts";
-export { BlockSuggestionRouter } from "./suggestion_router.ts";
+export { BlockSuggestionRouter } from "./block_suggestion_router.ts";
